Se pasa logica a viewmodel
